% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadKnitrCache.R
\name{loadKnitrCache}
\alias{loadKnitrCache}
\title{loadKnitrCache}
\usage{
loadKnitrCache(directory)
}
\arguments{
\item{directory}{A directory containing a knitr cache}
}
\value{
loads entire cache into global enviroment (lazy loading)
}
\description{
Loads a knitr cache into the current global enviroment
}
\details{
When a knitr script is run with cache enabled all data is saved in a given directory. However, this is not easily loaded into an active R session for easy manipulation when working on scripts.

This function allows you to load this data into the current R session. Simply call loadKnitrCache('directory/containing/cache/').

All data is 'lazyLoaded' meaning that R doesn't actually pull the data into memory until you begin working with it.
}
\examples{
loadKnitrCache('~/directory/containing/cache/')
}

